Add end-of-days message
We're shutting down Standup and want to tell the users.

We can't do a standup broadcast because it sends too much data all at once
and the IRC server kicks the bot. That's covered in some issue somewhere.

So this changes the reply message figuring that all users who use it via
IRC will see the reply message. Then we don't trigger broadcast issues
and we don't have to hunt people down.